Refactor custom components out of theme.h
